- added the Blob class which will be used for serialization - still need to plan the binary format used for serialization - few improvements overall

+++ src/crypto/xtea.h
@@ -2,7 +2,7 @@
 #define CRYPTO_XTEA_H_
 
 #include "../def.h"
-void xtea_encode(uint32 *k, uint8 *data, uint32 len);
-void xtea_decode(uint32 *k, uint8 *data, uint32 len);
+void xtea_encode(uint32 *k, uint8 *data, size_t len);
+void xtea_decode(uint32 *k, uint8 *data, size_t len);
 
 #endif //CRYPTO_XTEA_H_
